moved nav links data to navLinksData file
