+++ modules/shared/components/molecules/Navigation/INavigation.d.ts
@@ -1,6 +1,14 @@
+import type { ReactElement } from 'react';
+
 declare namespace INavigation {
   export interface IProps {
     profilePic?: string;
+  }
+
+  export interface INavLinks {
+    name: string;
+    path: string;
+    content: ReactElement;
   }
 }
 export { INavigation };
